Update mods example to use absolute paths
I found that using an absolute path in the `config_player.xml` for the mods works much more reliably. Neurotrauma wouldn't start at all for me without this change, due to an exception being thrown that involved what looked like an attempt to reference the relative path inside of the relative path given.